Overhauled tree folding and row toggle.
The previous implementation of the tree folding mechanism prevented row toggling for working on parent rows: either a row could support
toggling, or it could act as a subgroup parent, but not both. Toby wanted to retain the toggle feature for all rows, hence this
implementation that moves the toggle control from the first td in the row to a span inside that td, and adds an img element to handle
folding events.