Added noticeboard dispatcher script.
This is virtually a copy of megaphone's index.cgi, with some paths
changed to handle the different location of the resources. Otherwise it
is pretty much the same - DummyAuth has replaced SSHAuth to ensure that,
even if the user guesses a megaphone module id, they can't log in or use
it anyway.